Make synchronisation variables atomic.
Previously, the variables used for synchronisation between threads
during capture were declared as volatile. While this will work on some
platforms and compilers (it's usually safe on x86), it's not strictly
correct for the C++ memory model: you need to update the variable using
an operation with a defined ordering to ensure that other changes to
memory (e.g. data in the buffer) are visible to other threads.

Use atomics instead, with the default sequential-consistency semantics,
which should do the right thing on all platforms.