Run the capture thread with real-time priority.
This sets the scheduling class of the USB capture thread to SCHED_RR
while a capture is in progress. This should give the capture thread a
better chance of reacting quickly to transfers completing, especially if
there are other regular userspace processes competing with it.

If setting the priority fails (e.g. because the system ulimit for
real-time priority isn't high enough), it will print a warning and
continue as usual.